+++ src/services/security/index.ts
@@ -1,4 +1,9 @@
-﻿export { BackupEncryptionService } from "./BackupEncryptionService";
+export { BackupEncryptionService } from "./BackupEncryptionService";
+export {
+  CURRENT_BACKUP_PBKDF2_ITERATIONS,
+  LEGACY_BACKUP_PBKDF2_ITERATIONS,
+  isSupportedBackupPbkdf2Iterations,
+} from "./BackupEncryptionParameters";
 export { KeyDerivationService } from "./KeyDerivationService";
 export { SecureRandomService } from "./SecureRandomService";